% Generated by roxygen2 (4.0.2): do not edit by hand
\name{myqs}
\alias{myqs}
\title{Simple quantile calculator}
\usage{
myqs(x, at = c(0.01, 0.05, 0.1, 0.5, 0.9, 0.95, 0.99))
}
\arguments{
\item{x}{A vector for which quantiles are to be calculated}

\item{at}{A vector with p-values for the desired quantiles}
}
\value{
A named vector with the desired quantiles.
}
\description{
A simple demo function which returns quantiles
}
\details{
This is just an examples.
}
\examples{
set.seed(123)     # be reproducible
x <- rnorm(1000)
myqs(x)
}
\author{
Dirk Eddelbuettel
}
\references{
None
}
\seealso{
\link[stats:quantile]{quantile}
}

